Move AVR Wire library utility folder under source
