+++ hardware/arduino/avr/libraries/Wire/src/utility/twi.h
